requirements.txt
# Core Dependencies
PyYAML>=6.0
openai>=1.0.0
httpx>=0.24.0
psutil>=5.9.0

# Optional but Recommended
colorama>=0.4.6  # For colored terminal output

# MLX Dependencies (for Whisper models)
# mlx>=0.5.0  # Uncomment if using MLX models locally
# mlx-lm>=0.0.6  # Uncomment for MLX language models

# Audio Processing (if using VAD mode)
# torch>=2.0.0  # Uncomment for Voice Activity Detection
# torchaudio>=2.0.0  # Uncomment for VAD support

# Advanced UI (optional alternatives to tmux)
# rich>=13.0.0  # For rich terminal UI
# blessed>=1.20.0  # For terminal UI framework

# Web Interface (optional future enhancement)
# flask>=3.0.0  # For web server
# flask-socketio>=5.3.0  # For real-time updates
# python-socketio>=5.10.0  # WebSocket support